requirements.txt
openai
pandas
regex
numpy
scipy
gensim
scikit-learn
matplotlib
tiktoken
langchain
faiss-cpu
pypdf
llama-index
python-dotenv
numexpr
requests
langchain_community
langchain
openai